\name{Chau.boxplot}
\alias{Chau.boxplot}

\title{
Chauvenet-type boxplot
}

\description{
This is the Chauvenet-type boxplot.
}

\usage{
Chau.boxplot(x)
}

\arguments{
  \item{x}{The Chauvenet-type boxplot defines a new whisker compared to that for Tukey's boxplot.}
}

\author{
Tiejun Tong
}

