% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChauBoxplot-package.R
\docType{package}
\name{ChauBoxplot-package}
\alias{ChauBoxplot}
\alias{ChauBoxplot-package}
\title{ChauBoxplot: Chauvenet-type boxplot}
\description{
This package provides a modified boxplot with the new fence coefficient determined by Lin et al. (2025). Specifically, the fence coefficient k=1.5 in Tukey's boxplot has been replaced by the fence coefficient associated with Chauvenet’s criterion in their formula (9). For convenience, we have also provided the chauvenet-type boxplot in both base R with function chau_boxplot(), and in ggplot2 with function geom_chau_boxplot().
}
\references{
Hongmei Lin, Riquan Zhang and Tiejun Tong (2025). When Tukey meets Chauvenet: a new boxplot criterion for outlier detection. Journal of Computational and Graphical Statistics, accepted.
}
\author{
\strong{Maintainer}: Tiejun Tong \email{tongt@hkbu.edu.hk}

Authors:
\itemize{
  \item Hongmei Lin \email{hmlin@suibe.edu.cn}
  \item Riquan Zhang \email{rqzhang@suibe.edu.cn}
}

}
